moved utils.archive and utils.skyflats into robotic.utils, since they are pydantic models now
