+++ pyobs/modules/flatfield/scheduler.py
@@ -8,8 +8,8 @@ from pyobs.modules import Module
 from pyobs.modules import timeout
 from pyobs.object import get_object
 from pyobs.utils.parallel import event_wait
-from pyobs.utils.skyflats.priorities.base import SkyflatPriorities
-from pyobs.utils.skyflats.scheduler import Scheduler
+from pyobs.robotic.utils.skyflats.priorities.base import SkyflatPriorities
+from pyobs.robotic.utils.skyflats.scheduler import Scheduler
 from pyobs.utils.time import Time
 
 log = logging.getLogger(__name__)
