+++ pyobs/robotic/utils/archive/local_archive.py
@@ -8,7 +8,7 @@ from astropy.io import fits
 
 from pyobs.utils.time import Time
 from pyobs.images import Image
-from pyobs.utils.archive import Archive, FrameInfo
+from pyobs.robotic.utils.archive import Archive, FrameInfo
 from pyobs.utils.enums import ImageType
 
 log = logging.getLogger(__name__)
