+++ pyobs/robotic/utils/archive/pyobs_archive.py
@@ -7,7 +7,7 @@ from pydantic import PrivateAttr
 
 from pyobs.utils.time import Time
 from pyobs.images import Image
-from pyobs.utils.archive import Archive, FrameInfo
+from pyobs.robotic.utils.archive import Archive, FrameInfo
 from pyobs.utils.enums import ImageType
 
 log = logging.getLogger(__name__)
