+++ pyobs/robotic/utils/skyflats/pointing/base.py
