+++ pyobs/robotic/utils/skyflats/pointing/static.py
