+++ pyobs/robotic/utils/skyflats/priorities/base.py
