+++ pyobs/robotic/utils/skyflats/priorities/const.py
