+++ pyobs/utils/pipeline/night.py
@@ -6,7 +6,7 @@ from pyobs.object import get_object
 from pyobs.utils.time import Time
 from pyobs.utils.fits import FilenameFormatter
 from pyobs.images import Image
-from pyobs.utils.archive import Archive
+from pyobs.robotic.utils.archive import Archive
 from pyobs.utils.enums import ImageType
 from .pipeline import Pipeline
